Moved scrapper classes to lib
