+++ lib/api_scraper.js
@@ -1,0 +1,28 @@
+"use strict";
+
+const axios = require("axios");
+
+class ApiScraper {
+  constructor(url) {
+    this.url = url;
+    this.user_agent =
+      "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.132 Safari/537.36";
+    this.headers = {
+      headers: {
+        "User-Agent": this.user_agent,
+      },
+    };
+  }
+
+  async get() {
+    const response = await axios.get(this.url, this.headers);
+    return response.data;
+  }
+
+  async post(data) {
+    const response = await axios.post(this.url, data, this.headers);
+    return response.data;
+  }
+}
+
+module.exports = ApiScraper;
