+++ lib/index.js
@@ -1,0 +1,7 @@
+const Scraper = require("./scraper");
+const ApiScraper = require("./api_scraper");
+
+module.exports = {
+  Scraper,
+  ApiScraper,
+};
