+++ lib/scraper.js
@@ -1,0 +1,24 @@
+"use strict";
+
+const axios = require("axios");
+const Jssoup = require("jssoup").default;
+
+class Scraper {
+  constructor(url) {
+    this.url = url;
+    this.headers = {
+      headers: {
+        "User-Agent":
+          "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.132 Safari/537.36",
+      },
+    };
+    this.soup = this.get_soup();
+  }
+
+  async get_soup() {
+    const response = await axios.get(this.url, this.headers);
+    return new Jssoup(response.data);
+  }
+}
+
+module.exports = Scraper;
