+++ peviitor_scraper.js
@@ -2,47 +2,7 @@
 
 const axios = require("axios");
 const jssoup = require("jssoup").default;
-
-class Scraper {
-  constructor(url) {
-    this.url = url;
-    this.headers = {
-      headers: {
-        "User-Agent":
-          "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.132 Safari/537.36",
-      },
-    };
-    this.soup = this.get_soup();
-  }
-
-  async get_soup() {
-    const response = await axios.get(this.url, this.headers);
-    return new jssoup(response.data);
-  }
-}
-
-class ApiScraper {
-  constructor(url) {
-    this.url = url;
-    this.user_agent =
-      "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.132 Safari/537.36";
-    this.headers = {
-      headers: {
-        "User-Agent": this.user_agent,
-      },
-    };
-  }
-
-  async get() {
-    const response = await axios.get(this.url, this.headers);
-    return response.data;
-  }
-
-  async post(data) {
-    const response = await axios.post(this.url, data, this.headers);
-    return response.data;
-  }
-}
+const {Scraper, ApiScraper} = require("./lib");
 
 function postApiPeViitor(data, company, apikey = null) {
   const V4cleanUrl = "https://api.peviitor.ro/v4/clean/";
